% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laryngectcomy.R
\docType{data}
\name{laryngectomy}
\alias{laryngectomy}
\title{Predictors and outcomes for people underwent laryngectomy}
\format{
A tibble with 123 rows and 32 variables:
\describe{
\item{age}{Age in years with 90+ removed}
\item{gender}{sex (M=1, F=2)}
\item{recur_site}{Recurrent Site (Supra=1, Glottic=2, Transglottic=3, Unknown=4)}
\item{trt_mod}{Treatment Modality (XRT=1  Chemo+XRT=2 Unspecified=3)}
\item{int_stg}{Initial Stage  (T1=1, T2=2, T3=3, T4=4 Unknown=5)}
\item{recur_stg}{Recurrent staging (T1=1, T2=2, T3=3, T4=4, unknown=5)}
\item{preop_Img_mod}{Imaging (CT=1  PET=2  CT+PET=3 MRI=4)}
\item{nk_dissection}{Neck Dissection (Yes=1, No=2)}
\item{Laterality}{Laterality	(Bilateral Neck Dissection=1, Unilateral Neck Dissection=2)}
\item{mic_vas_flap}{Microvascular Flap (Yes=1, No=2)}
\item{pec_flap}{Pectoralis major myocutaneous flap (Yes=1, No=2)}
\item{pathology}{Nodal Pathology evaluation of specimen (Positive=1, Negative=2)}
\item{path_stg}{Pathologic staging (T1=1, T2=2, T3=3, T4=4, unknown=5)}
\item{complications}{Complications (Yes=1, No=2)}
\item{fistula}{Salivary fistula or leak (Yes=1, No=2)}
\item{wound_inf}{Wound Infection (Yes=1, No=2)}
\item{dehiscence}{Wound dehiscence (Yes=1, No=2)}
\item{wound_comp}{Wound Complications (Yes=1, No=2)}
\item{chyle_leak}{Chyle leak (Yes=1, No=2)}
\item{hematoma}{Hematoma (Yes=1, No=2)}
\item{revision_proc}{Revision procedure (Yes=1, No=2)}
\item{flap_fail}{Flap Failure (Yes=1, No=2)}
\item{periop_death}{Perioperative Death (Yes=1, No=2)}
\item{med_comp}{Medical Complications (Yes=1, No=2)}
\item{comp_2plus}{2 or more complications (Yes=1, No=2)}
\item{hosp_stay}{Hospital Stay (<7 days=1, 7-14 days=2, >14 days)}
\item{death}{Death status (Dead=1 Alive=2)}
\item{recur}{Recurrence status (Recurred=1, No recurrence=2)}
\item{sor}{Site of Recurrence after salvage laryngectomy (Locoregional=1, Distant=3, No recurrence=0)}
\item{length_fu}{Days of follow-up}
\item{death_fu}{Days until death}
\item{recur_fu}{Days until recurrence}
}
}
\usage{
data(laryngectomy)
}
\description{
A dataset containing demographics and surgery details and outcomes on 123
people who underwent laryngectomy at an unspecified medical center between
2000 and 2012.
}
\keyword{datasets}
